% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLinesToPlot.R
\name{addLinesToPlot}
\alias{addLinesToPlot}
\title{addLinesToPlot}
\usage{
addLinesToPlot(plot, yintercept = 0, xintercept = 0)
}
\arguments{
\item{plot}{The plot to add the lines to.}

\item{yintercept}{The y axis intercept for the horizontal line.}

\item{xintercept}{The x axis intercept for the vertical line.}
}
\value{
The plot with dashed gray lines as the x and y intercepts.
}
\description{
Adds a gray dashed line to the plot at the intercepts
}
