% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEnergypercentDensity.R
\name{calculateEnergypercentDensity}
\alias{calculateEnergypercentDensity}
\title{calculateEnergypercentDensity}
\usage{
calculateEnergypercentDensity(df)
}
\arguments{
\item{df}{A dataframe with of recipes with sample_id (recipe id), and names of the macronutrients like they are named in Matvaretabellen: Kilocalories, Kilojoules, Carbo, Sugar, Fat, SatFa, Protein, \verb{Dietary fibre}.}
}
\value{
A tidy dataframe with a sample_id, densityMJ and energy_percent column, and a feature column explaining which macronutrient the densityMJ/energy_perecent refers to.
}
\description{
Calculate percent of total energy from macronutrients, and density per MJ energy.
}
