% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNutritionScore_keyhole.R
\name{calculateNutritionScore_keyhole}
\alias{calculateNutritionScore_keyhole}
\title{calculateNutritionScore_keyhole}
\usage{
calculateNutritionScore_keyhole(df)
}
\arguments{
\item{df}{A long format dataframe with an Ingredients column, a Foodgroup column with food groups named as they are in analyses.R script, a sample_id column for recipe ID, and a feature column with nutrient names in line with Matvaretabellen, with their values in a value column.}
}
\value{
A dataframe with the recipe ID and a column noting if it is keyhole certified or not.
}
\description{
Score a recipe using the Keyhole system.
}
