% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNutritionScore_trafficlights.R
\name{calculateNutritionScore_trafficlights}
\alias{calculateNutritionScore_trafficlights}
\title{calculateNutritionScore_trafficlights}
\usage{
calculateNutritionScore_trafficlights(df, inverted = "yes")
}
\arguments{
\item{df}{A dataframe with a sample_id column for recipe ID, a salt column with salt in gram/100g recipe, and nutrient columns with names in line with Matvaretabellen.}

\item{inverted}{Should the returned score be inverted or not? Either 'yes' or 'no', default 'yes'.}
}
\value{
A dataframe with the recipe ID and its total traffic light score.
}
\description{
Score a recipe unsing the UK Traffic Light System.
}
