% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNutritionScore_who.R
\name{calculateNutritionScore_who}
\alias{calculateNutritionScore_who}
\title{calculateNutritionScore_who}
\usage{
calculateNutritionScore_who(df)
}
\arguments{
\item{df}{A dataframe with a sample_id column for recipe ID, and macronutrient names in line with Matvaretabellen.}
}
\value{
A dataframe with the recipe ID and its corresponding WHO score.
}
\description{
Score a recipe on how well it adheres to WHO macronutrient recommendations.
}
