% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeGGplotTxtSize.R
\name{changeGGplotTxtSize}
\alias{changeGGplotTxtSize}
\title{changeGGplotTxtSize}
\usage{
changeGGplotTxtSize(plot, txt_size = 12)
}
\arguments{
\item{plot}{A ggplot.}

\item{txt_size}{New text size to be used in plot, default is 12.}
}
\value{
A ggplot with resized text.
}
\description{
Change the text size of the axis title and text of a ggplot2 plot.
}
