% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectCorr.R
\name{collectCorr}
\alias{collectCorr}
\title{collectCorr}
\usage{
collectCorr(corr_test_results)
}
\arguments{
\item{corr_test_results}{The results dataframe from running corr.test.}
}
\value{
A dataframe with the correlation, its p-value and the level of significance in *.
}
\description{
Create a df with group, adjusted p.value and symbols for showing the significance of the p value from a corr.test result.
}
\details{
Helper function to get results from corr.test
}
