% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_manuals.R
\name{colorManuals}
\alias{colorManuals}
\title{colorManuals}
\usage{
colorManuals(df, group = TRUE, feature_anno = TRUE)
}
\arguments{
\item{df}{The dtaframe with dta to be plotted.}

\item{group}{Should a color manual be made for the 'group' column? Default TRUE.}

\item{feature_anno}{Should a color manual be made for the 'feature_anno' column? Default TRUE.}
}
\value{
A list with one or two named vectors with colors for the variables in the 'group' and/or feature_anno' columns.
}
\description{
Create a color manual to be used in various plots.
}
