% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPCA.R
\name{createPCA}
\alias{createPCA}
\title{createPCA}
\usage{
createPCA(
  df,
  principal_components = c(1, 2),
  plots = "scores",
  interesting_samples = NULL,
  cutoff = NULL
)
}
\arguments{
\item{df}{A tidy dataframe with at least sample_id, group, feature and value columns.}

\item{principal_components}{Principal components to be used as x and y axis, default is 1 and 2.}

\item{plots}{Which plots should be made? 'scores', 'loadings' or 'both'.}

\item{interesting_samples}{Character vector of samples to label in the score plot.}

\item{cutoff}{Numerical value used to label all features with a loading >cutoff or <-cutoff. Default NULL.}
}
\value{
A PCA loadings plot.
}
\description{
Creates a PCA loadings plot using ggplot.
}
