% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractStats.R
\name{extractStats}
\alias{extractStats}
\title{extractStats}
\usage{
extractStats(df, filterp = "yes", num = 0.05)
}
\arguments{
\item{df}{Rstatix results dataframe after Welch's t-test or Wilcoxon test.}

\item{filterp}{Should the dataframe be filtered for their p.adj value? Default 'yes'.}

\item{num}{If filterp is 'yes', only features with p.adj values < num are kept. Default is 0.05.}
}
\value{
A dataframe with the feature column and a new 'string' column with stats that can be used as titles for plots.
}
\description{
Get a string of stats for plots, from an rstatix-results df (t-test/wilcox + respective effect sizes)
}
