% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_manuals.R
\name{featureAnnotationColors}
\alias{featureAnnotationColors}
\title{featureAnnotationColors}
\usage{
featureAnnotationColors(df)
}
\arguments{
\item{df}{The dtaframe with dta to be plotted.}
}
\value{
A list with a named vector with colors for the variables in the 'feature_anno' column.
}
\description{
Create a color manual for the feature annotation factor to be used in plots.
}
