% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFeatures.R
\name{filterFeatures}
\alias{filterFeatures}
\title{filterFeatures}
\usage{
filterFeatures(df, features)
}
\arguments{
\item{df}{Dataframe to be filtered, must have a 'feature' column.}

\item{features}{Feature(s) to filter the dataframe by.}
}
\value{
A filtered dataframe.
}
\description{
Filter a 'feature' column in a dataframe for feature(s) of interest.
}
