% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterP.R
\name{filterP}
\alias{filterP}
\title{filterP}
\usage{
filterP(df, filterp = "yes", num = 0.05)
}
\arguments{
\item{df}{Dataframe to be filtered.}

\item{filterp}{Filter p.adj? 'yes' or 'no'? Default is 'yes'.}

\item{num}{p.adj below this number is kept. Default is 0.05.}
}
\value{
A filtered dataframe.
}
\description{
Filter a p.adj column in a dataframe for values below a certain number, default 0.05.
}
