% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTypeOfRecipe.R
\name{findTypeOfRecipe}
\alias{findTypeOfRecipe}
\title{findTypeOfRecipe}
\usage{
findTypeOfRecipe(df)
}
\arguments{
\item{df}{A dataframe with an column of recipe names titled "sample_id".}
}
\value{
The dataframe with a new column "recipe_type"
}
\description{
Decide type of recipe based on its name. Recipe type corresponds to Keyhole certification types of food.
}
\details{
What type of recipe is it?
}
