% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixRefID.R
\name{fixRefID}
\alias{fixRefID}
\title{fixRefID}
\usage{
fixRefID(reference, first_w, second_w = "\\\\")
}
\arguments{
\item{reference}{Reference database to look up.}

\item{first_w}{The correct first word in reference database.}

\item{second_w}{The correct second word in reference database.}
}
\value{
The dataframe with updated IDs.
}
\description{
Set a new ID number for an ingredient if checkRef made a mistake.
}
\details{
Fix the ID number of an ingredient if checkRef made a mistake.
}
