% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCAloadings.R
\name{getPCAloadings}
\alias{getPCAloadings}
\title{getPCAloadings}
\usage{
getPCAloadings(prcomp_object, principal_components)
}
\arguments{
\item{prcomp_object}{A prcomp object.}

\item{principal_components}{The principal components to extract the loadings from.}
}
\value{
A dataframe with the loadings of each feature from the principal components chosen.
}
\description{
Get the loading scores for each feature in a prcomp object from different principal components.
}
