% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCpercentages.R
\name{getPCpercentages}
\alias{getPCpercentages}
\title{getPCpercentages}
\usage{
getPCpercentages(prcomp_object)
}
\arguments{
\item{prcomp_object}{A prcomp object.}
}
\value{
A dataframe with the \% contributed by each principal component.
}
\description{
Takes a prcomp object and calculates the percentages each PC contributes.
}
