% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_manuals.R
\name{groupColors}
\alias{groupColors}
\title{groupColors}
\usage{
groupColors(df)
}
\arguments{
\item{df}{The dataframe with data to be plotted.}
}
\value{
A list with a named vector with colors for the variables in the 'group' column.
}
\description{
Create a color manual for the group factor to be used in plots.
}
