% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasWholeGrains.R
\name{hasWholeGrains}
\alias{hasWholeGrains}
\title{hasWholeGrains}
\usage{
hasWholeGrains(df)
}
\arguments{
\item{df}{A dataframe with recipes, with recipe names in a sample_id column, and ingredient names in rows in an Ingredient column.}
}
\value{
The dataframe with a new column "whole_grain_requirement", with "yes" if ingredients of recipe fulfill Keyhole whole grain criteria, "no" if not.
}
\description{
Checks if grain-based carbohydrates in a recipe fullfills criteria of Keyhole to be whole grains.
}
\details{
Does recipe fulfill whole grain criteria of the Keyhole certification?
}
