% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyholeCertification.R
\name{keyholeCertification}
\alias{keyholeCertification}
\title{keyholeCertification}
\usage{
keyholeCertification(df)
}
\arguments{
\item{df}{A dataframe with columns sample_id (recipe name), recipe_type, keyhole_amounts_pct, Sugar, Salt, SatFa, whole_grain_requirement}
}
\value{
The dataframe with a new column "keyhole_certified" with "Keyhole" or "No keyhole".
}
\description{
Checks if recipe fulfills criteria for Keyhole certification.
}
\details{
Does a recipe fullfill the Keyhole certification for its recipetype?
}
