% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCorrelations.R
\name{myCorrelations}
\alias{myCorrelations}
\title{myCorrelations}
\usage{
myCorrelations(data, mapping, ...)
}
\arguments{
\item{data}{Uses the data loaded from ggpairs.}

\item{mapping}{Uses the mapping from ggpairs.}

\item{...}{Uses arguments from ggpairs.}
}
\value{
The upper segment of ggpairs with correlation scores from corr.test, colored with custom colors.
}
\description{
Creates the upper part of ggpairs with correlation scores by group from corr.test, and groups can by customly colored.
}
\details{
Fixes the colors on the correlation scores in the upper segment of ggpairs, slight alteration of code from Isaac Zhao on statexchange
}
