% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAprep.R
\name{PCAprep}
\alias{PCAprep}
\title{PCAprep}
\usage{
PCAprep(df)
}
\arguments{
\item{df}{A tidy dataframe with sample_id, group, feature and value columns.}
}
\value{
A wide dataframe with subject_id and group affiliation as rownames
}
\description{
Turns a tidy dataframe into a wide dataframe that can be run through prcomp or ropls, keeping group affiliation of the samples.
}
