% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCAloadings.R
\name{plotPCAloadings}
\alias{plotPCAloadings}
\title{plotPCAloadings}
\usage{
plotPCAloadings(df, cutoff = NULL)
}
\arguments{
\item{df}{A dataframe with the principal components to be plotted in the first two columns.}

\item{cutoff}{Numerical value used to label all features with a loading >cutoff or <-cutoff. Default NULL.}
}
\value{
A PCA loadings plot.
}
\description{
Creates a PCA loadings plot using ggplot.
}
