% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCAscores.R
\name{plotPCAscores}
\alias{plotPCAscores}
\title{plotPCAscores}
\usage{
plotPCAscores(df, prc, interesting_samples = NULL)
}
\arguments{
\item{df}{Dataframe with the principal components to be plotted in the first two columns, and a group column to color by.}

\item{prc}{A vector with the percentage the principal components contribute, to be used for x and y axis titles.}

\item{interesting_samples}{Character vector of samples to label.}
}
\value{
A ggplot PCA score plot.
}
\description{
Creates a PCA score plot using ggplot.
}
