% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolinBox.R
\name{plotViolinBox}
\alias{plotViolinBox}
\title{plotViolinBox}
\usage{
plotViolinBox(df, r_statix = NULL)
}
\arguments{
\item{df}{Dataframe with a 'feature' column with feature names, a 'sample_id' column with sample ID's, a 'value' column with feature values for the samples, and a 'group' column splitting the samples into two groups (cases and controls).}

\item{r_statix}{An optional r statix t-test/wilcoxon test result table, that add a string of stats to the title of the plot.}
}
\value{
A half violin/half boxplot with individual data points.
}
\description{
Create a half violin/half boxplot with individual data points.
}
