% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseIngredients.R
\name{standardiseIngredients}
\alias{standardiseIngredients}
\title{standardiseIngredients}
\usage{
standardiseIngredients(df)
}
\arguments{
\item{df}{A dataframe with an Ingredients column, listing each ingredient of the recipe in individual rows.}
}
\value{
The dataframe with a new column "Ingredients_standardised".
}
\description{
Standardise ingredient names and adds them in a new column called "Ingredients_standardised".
}
\details{
Standardising ingredients names in a recipe
}
