% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseRecipes.R
\name{standardiseRecipes}
\alias{standardiseRecipes}
\title{standardiseRecipes}
\usage{
standardiseRecipes(df)
}
\arguments{
\item{df}{A dataframe with an Ingredient column with rowwise ingredients where amount + unit (i.e. 1 dl) is at the beginning of the row.}
}
\value{
The dataframe with all ingredient amounts and units in separate columns, with ingredient names and units  standardised.
}
\description{
Standardise all names of different ingredients, and extract to separate columns and standardise units of each ingredient, volume units to dl and weight units to kg Ingredients with similar density to water is turned into kg.
}
\details{
Standardise the ingredients in a recipe
}
