% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseUnits.R
\name{standardiseUnits}
\alias{standardiseUnits}
\title{standardiseUnits}
\usage{
standardiseUnits(df)
}
\arguments{
\item{df}{A dataframe with a unit column indicating what type of unit is used, and a numerical Amount column listing the amount.}
}
\value{
The dataframe with all ingredient units and volume standardised.
}
\description{
Standardise units, volume units to dl and weight units to grams.
}
\details{
Standardising the units in a recipe. Volume units to dl, weight to g.
}
