% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNutritionScore_nnr.R
\name{calculateNutritionScore_nnr}
\alias{calculateNutritionScore_nnr}
\title{calculateNutritionScore_nnr}
\usage{
calculateNutritionScore_nnr(df)
}
\arguments{
\item{df}{A dataframe with a sample_id column for recipe ID, and macronutrient names in line with Matvaretabellen.}
}
\value{
A dataframe with the recipe ID and its corresponding NNR score.
}
\description{
Score a recipe on how well it adheres to Nordic Nutritional Recommendations macronutrient recommendations.
}
