% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNutritionScore_nutriscore.R
\name{calculateNutritionScore_nutriscore}
\alias{calculateNutritionScore_nutriscore}
\title{calculateNutritionScore_trafficlights}
\usage{
calculateNutritionScore_nutriscore(df, inverted = "yes")
}
\arguments{
\item{df}{A dataframe with a sample_id column for recipe ID, a 'nutriscore_amounts_pct' column with the percentage of fruit/veg/legumes/nuts/healthy oils, and nutrient columns with names in line with Matvaretabellen (except kJ which is 'Kilojoules').}

\item{inverted}{Should the returned score be inverted or not? Either 'yes' or 'no', default 'yes'.}
}
\value{
A dataframe with the recipe ID and its nutriscore with corresponding letter.
}
\description{
Score a recipe unsing the UK Traffic Light System.
}
