% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRef.R
\name{checkRef}
\alias{checkRef}
\title{checkRef}
\usage{
checkRef(df, reference)
}
\arguments{
\item{df}{A dataframe with an Ingredients column to be checked against the reference database.}

\item{reference}{The reference database.}
}
\value{
A dataframe with the ingredient name and the reference ID of the first hit in the database. See references for how the databse must be built.
}
\description{
Checks if an ingredient is present in a reference database
}
\details{
Check if an ingredient can be found in a reference database.
}
